refactor: refactored old code
