+++ package.json
@@ -1,28 +1,29 @@
 {
-  "name": "toggl-track-mcp-server",
-  "version": "1.0.0",
-  "type": "module",
-  "private": true,
-  "description": "",
-  "main": "dist/index.js",
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "dev": "tsc -w -p tsconfig.json",
-    "start": "node dist/index.js",
-    "typecheck": "tsc -p tsconfig.json --noEmit",
-    "test": "echo \"No tests yet\""
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.18.0",
-    "dotenv": "^16.4.5",
-    "fastmcp": "^3.17.0",
-    "zod": "^3.25.76"
-  },
-  "devDependencies": {
-    "@types/node": "^24.3.3",
-    "typescript": "^5.9.2"
-  }
+	"name": "toggl-track-mcp-server",
+	"version": "1.0.0",
+	"type": "module",
+	"private": true,
+	"description": "",
+	"main": "dist/index.js",
+	"scripts": {
+		"build": "tsc -p tsconfig.json",
+		"dev": "tsc -w -p tsconfig.json",
+		"start": "node dist/index.js",
+		"typecheck": "tsc -p tsconfig.json --noEmit",
+		"test": "echo \"No tests yet\""
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"dependencies": {
+		"@modelcontextprotocol/sdk": "^1.18.0",
+		"dotenv": "^16.4.5",
+		"fastmcp": "^3.17.0",
+		"zod": "^3.25.76"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "2.2.4",
+		"@types/node": "^24.3.3",
+		"typescript": "^5.9.2"
+	}
 }
