+++ src/schemas/date.ts
@@ -1,0 +1,2 @@
+export type Timezone = "UTC" | "Asia/Tokyo";
+export type ISODateString = string; // YYYY-MM-DD
