+++ src/tools/ping.ts
@@ -1,0 +1,8 @@
+export const pingTool = {
+	name: "ping",
+	description: 'Health check tool. Returns "ok".',
+	annotations: { readOnlyHint: true, idempotentHint: true, title: "Ping" },
+	async execute() {
+		return "ok";
+	},
+};
